data_entry_from_image_file/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-wtf>=1.2.2",
    "flask-pymongo>=3.0.1",
    "google-generativeai>=0.8.4",
    "werkzeug>=3.1.3",
    "trafilatura>=2.0.0",
    "sqlalchemy>=2.0.38",
    "pillow>=11.1.0",
    "spacy>=3.8.4",
    "numpy>=2.2.3",
    "pytesseract>=0.3.13",
]
